Adicionando CRUD para Usuário e implementando camadas de Service, Repository, DTO e Controller; Criando DadoInexistenteException e lançando-a no lugar de DadoExistenteException nas classe de Service; Criando converter para Usuário;

+++ src/main/java/br/com/inteligentclin/service/UsuarioService.java
@@ -1,12 +1,18 @@
 package br.com.inteligentclin.service;
 
+import br.com.inteligentclin.dtos.converters.UsuarioModelMapperConverter;
+import br.com.inteligentclin.dtos.usuarioDTO.UsuarioModelDTO;
+import br.com.inteligentclin.dtos.usuarioDTO.UsuarioSummaryDTO;
 import br.com.inteligentclin.entity.Usuario;
 import br.com.inteligentclin.repository.IUsuarioRepository;
+import br.com.inteligentclin.service.exception.DadoExistenteException;
+import br.com.inteligentclin.service.exception.DadoInexistenteException;
+import br.com.inteligentclin.service.exception.EntidadeRelacionadaException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -15,25 +21,54 @@ public class UsuarioService {
     @Autowired
     private IUsuarioRepository usuarioRepository;
 
-    public Usuario salvar(Usuario usuario) {
-        return usuarioRepository.save(usuario);
+    @Autowired
+    private UsuarioModelMapperConverter usuarioConverter;
+
+    public UsuarioModelDTO salvar(UsuarioModelDTO usuarioDTO) throws DadoExistenteException {
+        Optional<Usuario> usuarioBase = usuarioRepository.findByLogin(usuarioDTO.getLogin());
+        if (usuarioBase.isPresent())
+            throw new DadoExistenteException("O login: " + usuarioBase.get().getLogin() + " já existe. Tente outro " +
+                    "diferente.");
+        Usuario usuario = usuarioConverter.mapModelDTOToEntity(usuarioDTO, Usuario.class);
+        Usuario usuarioSalvo = usuarioRepository.saveAndFlush(usuario);
+        return usuarioConverter.mapEntityToModelDTO(usuarioSalvo, UsuarioModelDTO.class);
     }
 
-    public Optional<Usuario> buscarPorId(Long id) {
-        return usuarioRepository.findById(id);
+    public Optional<UsuarioModelDTO> buscarPorId(Long id) {
+        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new DadoInexistenteException("" +
+                "O usuário informado não pode ser localizado na base de dados"));
+
+        return Optional.ofNullable(usuarioConverter.mapEntityToModelDTO(usuario, UsuarioModelDTO.class));
     }
 
-    public List<Usuario> buscarTodos() {
-        return usuarioRepository.findAll();
+    public Page<UsuarioSummaryDTO> buscarTodos(Pageable pageable) {
+        return usuarioRepository.findAll(pageable)
+                .map(usuario -> usuarioConverter.mapEntityToSummaryDTO(usuario, UsuarioSummaryDTO.class));
     }
 
-    public void excluirPorId(Long id) {
-        try {
-            usuarioRepository.deleteById(id);
-        }catch (DataIntegrityViolationException e) {
-            throw new DataIntegrityViolationException("Não é possível excluir um usuário que está vinculado a uma " +
-                    "consulta.");
-        }
+    public void excluirPorId(Long id) throws EntidadeRelacionadaException {
+        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new DadoInexistenteException("" +
+                "O usuário informado não foi localizado na base de dados."));
+        boolean temConsulta = !usuario.getConsultas().isEmpty();
+        if (temConsulta)
+            throw new EntidadeRelacionadaException("Não é possível excluir o usuário de login: " + usuario.getLogin() +
+                    " pois está vinculado a uma ou mais consultas.");
+        usuarioRepository.deleteById(id);
+    }
+
+    public void atualizar(Long id, UsuarioModelDTO usuarioDTO) {
+
+        Usuario usuarioDaBase = usuarioRepository.findById(id).orElseThrow(() ->
+                new DadoInexistenteException("Usuário não encontrado na base de dados.")
+        );
+        usuarioDaBase.setNome(usuarioDTO.getNome());
+        usuarioDaBase.setSobrenome(usuarioDTO.getSobrenome());
+        usuarioDaBase.setCpf(usuarioDTO.getCpf());
+        usuarioDaBase.setEmail(usuarioDTO.getEmail());
+        usuarioDaBase.setTelefone(usuarioDTO.getTelefone());
+        usuarioDaBase.setSenha(usuarioDTO.getSenha());
+        usuarioDaBase.setCargo(usuarioDTO.getCargo());
+        usuarioRepository.save(usuarioDaBase);
     }
 
 }
